% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run.R
\name{app_run}
\alias{app_run}
\title{Run shiny application}
\usage{
app_run(.port = 3838, .host = "127.0.0.1")
}
\arguments{
\item{.port}{port number on which the application listens. Default is 3838.}

\item{.host}{host address where the application runs. Default is 127.0.0.1.}
}
\value{
a shiny.appobj object, run the app and keep listening in a port
}
\description{
Runs the shiny app, keeping listening in a given port (by default 3838).
The default host address is 127.0.0.1 to be run by default as localhost for
security reasons. It may be changed to 0.0.0.0, which allows the app to
listen on all available network interfaces.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
app_run()
\dontshow{\}) # examplesIf}
}
