% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Server-side app code}
\usage{
app_server(input, output)
}
\arguments{
\item{input}{A Shiny input object.}

\item{output}{A Shiny output object.}
}
\value{
A server object, used as the server argument in a call to
shiny::shinyApp()
}
\description{
Internal function, called within app_run()
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
shiny::shinyApp(app_ui(), app_server)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=app_ui]{app_ui()}}
}
