% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{App user interface}
\usage{
app_ui(request)
}
\arguments{
\item{request}{A shiny request object.}
}
\value{
A bslib_page object, used as the UI argument in a call to
shiny::shinyApp()
}
\description{
#' Internal function, called within app_run()
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
shiny::shinyApp(app_ui(), app_server)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=app_server]{app_server()}}
}
