% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ui.R
\name{mytheme}
\alias{mytheme}
\title{Create a custom Bootstrap theme}
\usage{
mytheme()
}
\value{
A Bootstrap theme object with customized colors and fonts.
}
\description{
This function generates a Bootstrap theme using a custom color palette.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mytheme()
\dontshow{\}) # examplesIf}
}
