% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.shiny.app-package.R
\docType{package}
\name{template.shiny.app-package}
\alias{template.shiny.app}
\alias{template.shiny.app-package}
\title{template.shiny.app: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
